+++ docker/services/ruby/apps/language_practice_plus/language_practice_plus_app.rb
@@ -44,6 +44,7 @@ class LanguagePracticePlus < MonadicApp
       "description": description,
       "icon": icon,
       "initiate_from_assistant": true,
+      "image": true,
       "pdf": false
     }
   end