+++ docker/services/ruby/apps/translate/translate_app.rb
@@ -28,6 +28,7 @@ class Translate < MonadicApp
       "description": description,
       "icon": icon,
       "initiate_from_assistant": true,
+      "image": true,
       "pdf": false
     }
   end