+++ docker/services/ruby/apps/voice_chat/voice_chat_app.rb
@@ -33,6 +33,7 @@ class VoiceChat < MonadicApp
       "icon": icon,
       "description": description,
       "initiate_from_assistant": true,
+      "image": true,
       "pdf": false
     }
   end