+++ docker/services/ruby/apps/voice_interpreter/voice_interpreter_app.rb
@@ -41,6 +41,7 @@ class VoiceInterpreter < MonadicApp
       "description": description,
       "icon": icon,
       "initiate_from_assistant": true,
+      "image": true,
       "pdf": false
     }
   end