+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### Change Log
 
-- [June, 2024] Automatic image rebuild feature
+- [June, 2024] Monadic Chat Console UI improved
+- [June, 2024] AI User feature introduced
+- [June, 2024] Automatic Docker image rebuild feature
 - [June, 2024] Source code block copy button added
 - [June, 2024] Video Describer app added
 - [May, 2024] Talk to Claude/Command R/Gemini apps added